+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 ## 2025-09-30
 * update to version 0.3.3 of [the crt shaders](https://github.com/jimhigson/jims_shaders.git)
 
+## 2025-09-27
+
+The game got pretty weird if one character (Head or Heels) lost all their lives. What is the other one supposed to do?
+Adopted the Retrospect remake model:
+
+* If one character loses all lives, and other has >2, other sacrifices 2, to give 1 to the character who lost
+* If one character loses all lives, and other has exactly 2, other sacrifices 1, to give 1 to the character who lost (continue with 1 each)
+* If other has 1, character goes to zero and is out of the game
+
 ## 2025-09-11
 * user can choose game speed from the menu (1x, 1.2x, 1.5x, 2x)
 * 1.2x is now the default game speed